+++ crates/miden-lib/src/testing/mock_account.rs
@@ -70,7 +70,7 @@ pub trait MockAccountExt {
 
         let asset = NonFungibleAsset::mock(&constants::NON_FUNGIBLE_ASSET_DATA_2);
         let non_fungible_storage_map =
-            StorageMap::with_entries([(asset.vault_key(), asset.into())]).unwrap();
+            StorageMap::with_entries([(asset.vault_key().into(), asset.into())]).unwrap();
         let storage =
             AccountStorage::new(vec![StorageSlot::Map(non_fungible_storage_map)]).unwrap();