+++ crates/miden-objects/src/asset/vault/partial.rs
@@ -2,7 +2,7 @@ use alloc::string::ToString;
 
 use miden_crypto::merkle::{InnerNodeInfo, MerkleError, PartialSmt, SmtLeaf, SmtProof};
 
-use super::AssetVault;
+use super::{AssetVault, VaultKey};
 use crate::Word;
 use crate::asset::{Asset, AssetWitness};
 use crate::errors::PartialAssetVaultError;
@@ -60,7 +60,7 @@ impl PartialVault {
         // vault. This is the most minimal and correct partial vault we can build.
         // TODO: Workaround for https://github.com/0xMiden/miden-base/issues/1966. Fix when implemented.
         partial_vault
-            .add(vault.open(Word::empty()))
+            .add(vault.open(VaultKey::new_unchecked(Word::empty())))
             .expect("adding the first proof should never fail");
 
         partial_vault
@@ -97,10 +97,10 @@ impl PartialVault {
     ///
     /// Returns an error if:
     /// - the key is not tracked by this partial vault.
-    pub fn open(&self, vault_key: Word) -> Result<AssetWitness, PartialAssetVaultError> {
+    pub fn open(&self, vault_key: VaultKey) -> Result<AssetWitness, PartialAssetVaultError> {
         let smt_proof = self
             .partial_smt
-            .open(&vault_key)
+            .open(&vault_key.into())
             .map_err(PartialAssetVaultError::UntrackedAsset)?;
         // SAFETY: The partial vault should only contain valid assets.
         Ok(AssetWitness::new_unchecked(smt_proof))
@@ -114,8 +114,8 @@ impl PartialVault {
     ///
     /// Returns an error if:
     /// - the key is not tracked by this partial SMT.
-    pub fn get(&self, vault_key: Word) -> Result<Option<Asset>, MerkleError> {
-        self.partial_smt.get_value(&vault_key).map(|word| {
+    pub fn get(&self, vault_key: VaultKey) -> Result<Option<Asset>, MerkleError> {
+        self.partial_smt.get_value(&vault_key.into()).map(|word| {
             if word.is_empty() {
                 None
             } else {
@@ -158,7 +158,7 @@ impl PartialVault {
                 PartialAssetVaultError::InvalidAssetInSmt { entry: *asset, source }
             })?;
 
-            if asset.vault_key() != *vault_key {
+            if *vault_key != asset.vault_key().into() {
                 return Err(PartialAssetVaultError::VaultKeyMismatch {
                     expected: asset.vault_key(),
                     actual: *vault_key,