+++ crates/miden-testing/src/kernel_tests/tx/test_prologue.rs
@@ -685,7 +685,7 @@ pub async fn create_account_non_fungible_faucet_invalid_initial_reserved_slot() 
     // Create a storage map with a mock asset to make it non-empty.
     let asset = NonFungibleAsset::mock(&[1, 2, 3, 4]);
     let non_fungible_storage_map =
-        StorageMap::with_entries([(asset.vault_key(), asset.into())]).unwrap();
+        StorageMap::with_entries([(asset.vault_key().into(), asset.into())]).unwrap();
     let storage = AccountStorage::new(vec![StorageSlot::Map(non_fungible_storage_map)]).unwrap();
 
     let account = AccountBuilder::new([1; 32])