+++ crates/miden-testing/src/tx_context/context.rs
@@ -7,7 +7,7 @@ use miden_lib::transaction::TransactionKernel;
 use miden_objects::account::{Account, AccountId, PartialAccount, StorageMapWitness, StorageSlot};
 use miden_objects::assembly::debuginfo::{SourceLanguage, Uri};
 use miden_objects::assembly::{SourceManager, SourceManagerSync};
-use miden_objects::asset::AssetWitness;
+use miden_objects::asset::{AssetWitness, VaultKey};
 use miden_objects::block::{AccountWitness, BlockHeader, BlockNumber};
 use miden_objects::note::{Note, NoteScript};
 use miden_objects::transaction::{
@@ -230,7 +230,7 @@ impl DataStore for TransactionContext {
         &self,
         account_id: AccountId,
         vault_root: Word,
-        vault_key: Word,
+        asset_key: VaultKey,
     ) -> impl FutureMaybeSend<Result<AssetWitness, DataStoreError>> {
         async move {
             if account_id == self.account().id() {
@@ -241,7 +241,7 @@ impl DataStore for TransactionContext {
                     )));
                 }
 
-                Ok(self.account().vault().open(vault_key))
+                Ok(self.account().vault().open(asset_key))
             } else {
                 let (foreign_account, _witness) = self
                     .foreign_account_inputs
@@ -264,7 +264,7 @@ impl DataStore for TransactionContext {
                     )));
                 }
 
-                Ok(foreign_account.vault().open(vault_key))
+                Ok(foreign_account.vault().open(asset_key))
             }
         }
     }