+++ crates/miden-tx/src/errors/mod.rs
@@ -6,6 +6,7 @@ use core::error::Error;
 use miden_lib::transaction::TransactionAdviceMapMismatch;
 use miden_objects::account::AccountId;
 use miden_objects::assembly::diagnostics::reporting::PrintDiagnostic;
+use miden_objects::asset::VaultKey;
 use miden_objects::block::BlockNumber;
 use miden_objects::crypto::merkle::SmtProofError;
 use miden_objects::note::{NoteId, NoteMetadata};
@@ -287,11 +288,11 @@ pub enum TransactionKernelError {
         source: DataStoreError,
     },
     #[error(
-        "failed to get vault asset witness from data store for vault root {vault_root} and vault_key {vault_key}"
+        "failed to get vault asset witness from data store for vault root {vault_root} and vault_key {asset_key}"
     )]
     GetVaultAssetWitness {
         vault_root: Word,
-        vault_key: Word,
+        asset_key: VaultKey,
         // thiserror will return this when calling Error::source on TransactionKernelError.
         source: DataStoreError,
     },